Improved ui code layout
